workers/requirements.txt
# Python worker dependencies
# Install with: pip install -r requirements.txt

# PostgreSQL async driver
asyncpg==0.29.0
# Sync DB access for simple scripts
psycopg2-binary==2.9.9

# HTTP client — httpx is async-first, with sync support
httpx==0.27.0

# Environment variable loading
python-dotenv==1.0.1

# Data validation (mirrors Zod on the TS side)
pydantic==2.7.1

# Structured logging
structlog==24.1.0
